% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{brapi_get_variants}
\alias{brapi_get_variants}
\title{BrAPI Get Geno Map}
\usage{
brapi_get_variants(start = NULL, end = NULL, chrom = NULL, simplify = TRUE)
}
\arguments{
\item{start}{start}

\item{end}{end}

\item{chrom}{chrom}

\item{simplify}{simplify}
}
\description{
BrAPI Get Geno Map
}
\examples{
if (interactive()) {
  # Configure your GIGWA connection
  set_qbms_config("https://gigwa.southgreen.fr/gigwa/", 
                  time_out = 300, engine = "gigwa", no_auth = TRUE)

  # Select a database by name
  gigwa_set_db("Sorghum-JGI_v1")

  # Select a project by name
  gigwa_set_project("Nelson_et_al_2011")
  
  # Select a specific run by name
  gigwa_set_run("run1")
  
  geno_map <- brapi_get_variants(start = 0, end = 12345678, chrom = c("Sb07"))
}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
