% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{brapi_headers}
\alias{brapi_headers}
\title{Common HTTP Headers to Send}
\usage{
brapi_headers()
}
\value{
A list of common HTTP headers to send.
}
\description{
Builds the list of common HTTP headers to send with each API call.
}
