% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\docType{data}
\name{brapi_map}
\alias{brapi_map}
\title{/brapi/{brapi_ver}/{brapi_call}}
\format{
An object of class \code{data.frame} with 25 rows and 3 columns.

An object of class \code{data.frame} with 25 rows and 3 columns.

An object of class \code{data.frame} with 25 rows and 3 columns.

An object of class \code{data.frame} with 25 rows and 3 columns.

An object of class \code{data.frame} with 25 rows and 3 columns.

An object of class \code{data.frame} with 25 rows and 3 columns.

An object of class \code{data.frame} with 25 rows and 3 columns.

An object of class \code{data.frame} with 25 rows and 3 columns.
}
\usage{
brapi_map

brapi_map

brapi_map

brapi_map

brapi_map

brapi_map

brapi_map

brapi_map
}
\description{
/brapi/{brapi_ver}/{brapi_call}

brapi v1 use get_program_trials which trialDbId={trialDbId}

POST: germplasmDbIds, observationLevel = "PLOT"

to get studyDbId

POST: studyDbIds

POST: studyDbIds (to get variantSetDbId)

POST: studyDbIds

POST: germplasmDbIds (use gigwa_get_samples call to get germplasmDbIds)
}
\keyword{datasets}
