% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{brapi_post_search_call}
\alias{brapi_post_search_call}
\title{Internal Function Used for Core BrAPI POST Calls}
\usage{
brapi_post_search_call(call_url, call_body, nested = TRUE)
}
\arguments{
\item{call_url}{BrAPI URL for the POST request.}

\item{call_body}{The request body to send with the POST request.}

\item{nested}{Logical indicating whether to flatten the nested structure. Default is TRUE.}
}
\value{
A list of results obtained from the BrAPI POST call.
}
\description{
This function is used internally to execute POST calls to BrAPI endpoints and
retrieve the results while handling pagination and long-running tasks.
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
