% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{debug_qbms}
\alias{debug_qbms}
\title{Debug internal QBMS status object}
\usage{
debug_qbms()
}
\value{
an environment object for the package config and status
}
\description{
Return the internal QBMS status object for debugging
}
\examples{
obj <- debug_qbms()
obj$config
obj$state
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
