% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{engine_pre_process}
\alias{engine_pre_process}
\title{NOTE: search wherever we have "qbms_globals$config$engine" condition in qbms.R}
\usage{
engine_pre_process(call_url, engine, func_name)
}
\description{
NOTE: search wherever we have "qbms_globals$config$engine" condition in qbms.R
}
