% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{germplasm_get_metadata}
\alias{germplasm_get_metadata}
\title{Get the Metadata of the Current Germplasm List}
\usage{
germplasm_get_metadata()
}
\value{
A data.frame of all metadata associated with the samples in the selected run.
}
\description{
This function retrieves the metadata of the current active run
as configured in the internal state object using the `genotyping_set_run()` function.
}
\examples{
if (interactive()) {
  # Configure your GIGWA connection
  set_qbms_config("https://gigwa.southgreen.fr/gigwa/", 
                  time_out = 300, engine = "gigwa", no_auth = TRUE)

  # Select a database by name
  gigwa_set_db("3kG_10M")

  # Select a project by name
  gigwa_set_project("3003_ind")
  
  # Select a specific run by name
  genotyping_set_run("1")
  
  # Get a list of all samples in the selected run
  metadata <- germplasm_get_metadata()
}
}
\seealso{
\code{\link{set_qbms_config}}, \code{\link{genotyping_set_run}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
