% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{get_auth_step_type}
\alias{get_auth_step_type}
\title{Get Authentication Step Type}
\usage{
get_auth_step_type(engine, no_auth)
}
\arguments{
\item{engine}{Character string: one of "bms", "breedbase", "deltabreed", "ebs", "germinate", "gigwa".}

\item{no_auth}{Logical. If TRUE, authentication is skipped.}
}
\value{
Character string: "none" if no_auth is TRUE, "token" if engine is
  "ebs" and no_auth is FALSE, or "credentials" for all other engines when
  no_auth is FALSE.
}
\description{
Determines the type of authentication step required for a given engine and
no_auth combination.
}
\examples{
if (interactive()) {
  get_auth_step_type("bms", FALSE)
  # [1] "credentials"

  get_auth_step_type("ebs", FALSE)
  # [1] "token"

  get_auth_step_type("bms", TRUE)
  # [1] "none"
}
}
