% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{get_climate}
\alias{get_climate}
\title{Get TerraClimate data for a given location}
\usage{
get_climate(
  lat,
  lon,
  from = "1958-01-01",
  to = "2020-12-31",
  clim_vars = NULL,
  month_mask = NULL
)
}
\arguments{
\item{lat}{Latitude in decimal degree format.}

\item{lon}{Longitude in decimal degree format.}

\item{from}{Start date as a string in the 'YYYY-MM-DD' format.}

\item{to}{End date as a string in the 'YYYY-MM-DD' format.}

\item{clim_vars}{A list of all climate variables to be imported. Valid list includes: aet, def, pet,
ppt, q, soil, srad, swe, tmax, tmin, vap, ws, vpd, and PDSI. Default is NULL for all.}

\item{month_mask}{A list of all months of interest (e.g., planting season: c(10:12,1:5)). 
Default is NULL for all.}
}
\value{
A list of two data.frame(s):
        - climate: includes the climate variables (https://www.climatologylab.org/terraclimate-variables.html).
        - biovars: includes the calculated bioclimatic variables (https://www.worldclim.org/data/bioclim.html).
}
\description{
Scientific Paper:    https://www.nature.com/articles/sdata2017191
About the Dataset:   https://www.climatologylab.org/terraclimate.html
Hosting Catalog:     http://thredds.northwestknowledge.net:8080/thredds/terraclimate_aggregated.html
Service Provider:    https://hpc.uidaho.edu/
Climate Variables:   https://www.climatologylab.org/terraclimate-variables.html
Bioclimatic Vars:    https://www.worldclim.org/data/bioclim.html
Calculating biovars: https://github.com/rspatial/dismo/blob/master/R/biovars.R

TerraClimate vs. WorldClim: https://www.worldclim.org/data/worldclim21.html
1958-2020 vs. 1970-2000
14 vs. 7 climate variables
~4 km vs. ~1 km spatial resolution
need to calculate vs. pre-calculated 19 bioclimatic variables
}
\examples{
if(interactive()) {
# data <- get_climate(36.016, 36.943, '1979-09-01', '2012-06-30', c('ppt', 'tmin', 'tmax'), c(10:12,1:5))
data <- get_climate(36.016, 36.943, '1979-09-01', '2012-06-30')

View(data$climate)

View(data$biovars)
}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
