% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{get_germplasm}
\alias{get_germplasm}
\title{Get all germplasm for a given crop}
\usage{
get_germplasm(crop_name = "bean")
}
\arguments{
\item{crop_name}{the name of the crop}
}
\value{
data.frame with all germplasm information
}
\description{
Get all germplasm for a given crop
}
\examples{
# In progress
}
