% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{get_germplasm_attributes}
\alias{get_germplasm_attributes}
\title{Get germplasm attributes for a given germplasmDbId in a crop}
\usage{
get_germplasm_attributes(
  crop_name = "bean",
  germplasmDbId = "49f77560-7874-11eb-91c9-0242ac140003"
)
}
\arguments{
\item{crop_name}{the name of the crop}

\item{germplasmDbId}{id in the database}
}
\value{
data.frame with attributes
}
\description{
Get germplasm attributes for a given germplasmDbId in a crop
}
\examples{
# In progress
}
