% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{get_login_details}
\alias{get_login_details}
\title{Login pop-up window}
\usage{
get_login_details()
}
\value{
a vector of inserted username and password
}
\description{
Build a GUI pop-up window using Tcl/Tk to insert BMS username and password
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
