% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gigwa.R
\name{get_marker_map}
\alias{get_marker_map}
\title{Get marker position information}
\usage{
get_marker_map()
}
\value{
A data frame with three columns:
\describe{
  \item{rs#}{Marker name (variant name)}
  \item{chrom}{Chromosome or linkage group name}
  \item{pos}{Genetic or physical position of the marker}
}
}
\description{
Retrieves the genetic marker map associated with the currently active study.
This includes the marker name, chromosome (or linkage group), and position.
}
\seealso{
\code{\link{set_qbms_config}}, \code{\link{set_study}}
}
\author{
Khaled Al-Shamaa (\email{k.el-shamaa@cgiar.org})
}
