% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{get_program_trials}
\alias{get_program_trials}
\title{Retrieve the List of Trials for the Active Breeding Program}
\usage{
get_program_trials()
}
\value{
A data frame containing information on trials for the active breeding program.
}
\description{
This internal function retrieves the list of trials for the currently active breeding program 
and crop combination. The crop and program must be set using \code{\link{set_crop}} and 
\code{\link{set_program}} prior to calling this function.
}
\seealso{
\code{\link{login_bms}}, \code{\link{set_crop}}, \code{\link{set_program}}, \code{\link{list_trials}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
