% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{get_trial_obs_ontology}
\alias{get_trial_obs_ontology}
\title{Get the traits ontology/metadata of the current active trial}
\usage{
get_trial_obs_ontology()
}
\value{
a data frame of the traits ontology/metadata
}
\description{
This function will retrieve the traits ontology/metadata of the current active
trial as configured in the internal state object using `set_trial()` function.
}
\examples{
if(interactive()) {
# config your server connection
set_qbms_config("https://bms.icarda.org/ibpworkbench")

# login using your account (interactive mode)
# you can pass your username and password as parameters (batch mode)
login_bms()

set_crop("wheat")

# select a breeding program by name
set_program("Wheat International Nurseries")

# select a specific study/trial by name
set_trial("IDYT39")

# get observation variable ontology
ontology <- get_trial_obs_ontology()
}
}
\seealso{
\code{\link{login_bms}}, \code{\link{set_crop}}, \code{\link{set_program}}, \code{\link{set_trial}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
