% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{get_trial_pedigree}
\alias{get_trial_pedigree}
\title{Get the Pedigree table for the Selected Trial}
\usage{
get_trial_pedigree()
}
\value{
A data frame with germplasm and parent details for the selected trial.
}
\description{
Get the pedigree table representing the pedigree tree for the currently active 
trial. Each row corresponds to a germplasm entry and includes identifiers and 
details for both parents.
}
\seealso{
\code{\link{login}}, \code{\link{set_crop}}, \code{\link{set_program}}, \code{\link{set_trial}} for related operations on crops, programs, and trials.
}
\author{
Khaled Al-Shamaa (\email{k.el-shamaa@cgiar.org})
}
