% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gigwa.R
\name{get_variants}
\alias{get_variants}
\title{Get Marker Matrix from the Selected Variant Set}
\usage{
get_variants()
}
\description{
Get a two-dimensional marker matrix for all samples in the currently active 
variant set. This function is used as an alternative to the \code{get_variantset()} 
function when the server does not support it.

Note: This approach is significantly slower than \code{\link{get_variantset}} 
because it fetches data page by page through the BrAPI API, rather than 
downloading the full matrix in one Falpjack format file.
}
\seealso{
\code{\link{set_qbms_config}}, \code{\link{set_variantset}}
}
\author{
Khaled Al-Shamaa (\email{k.el-shamaa@cgiar.org})
}
