% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gigwa.R
\name{get_variantset}
\alias{get_variantset}
\title{Get Marker Matrix from the Selected Variant Set}
\usage{
get_variantset()
}
\value{
A data frame of genotype calls in character format, with markers as rows and 
samples as columns.
}
\description{
Downloads a two-dimensional marker matrix for all samples in the currently 
active variant set, using the efficient Flapjack file format from the server. 
Row names correspond to marker IDs and column names correspond to unique 
sample IDs. Genotype calls are returned in character format.
}
\seealso{
\code{\link{set_qbms_config}}, \code{\link{set_variantset}}
}
\author{
Khaled Al-Shamaa (\email{k.el-shamaa@cgiar.org})
}
