% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{get_workflow_key}
\alias{get_workflow_key}
\title{Get Engine-to-Workflow Mapping}
\usage{
get_workflow_key(engine)
}
\arguments{
\item{engine}{Character string: one of "bms", "breedbase", "deltabreed", "ebs", "germinate", "gigwa".}
}
\value{
Character string: "standard", "ebs", or "gigwa".
}
\description{
Returns the workflow key for a given engine name.
}
\keyword{internal}
