% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gigwa.R
\name{gigwa_get_allelematrix}
\alias{gigwa_get_allelematrix}
\title{Get GIGWA Marker Matrix}
\usage{
gigwa_get_allelematrix(
  samples = NULL,
  start = 0,
  end = "",
  chrom = NULL,
  snps = NULL,
  snps_pageSize = 10000,
  samples_pageSize = 100,
  simplify = TRUE
)
}
\arguments{
\item{samples}{samples}

\item{start}{start}

\item{end}{end}

\item{chrom}{chrom}

\item{snps}{snps}

\item{snps_pageSize}{snps_pageSize}

\item{samples_pageSize}{samples_pageSize}

\item{simplify}{simplify}
}
\description{
Get GIGWA Marker Matrix
}
\examples{
if (interactive()) {
  # Configure your GIGWA connection
  set_qbms_config("https://gigwa.southgreen.fr/gigwa/", 
                  time_out = 300, engine = "gigwa", no_auth = TRUE)

  # Select a database by name
  gigwa_set_db("Sorghum-JGI_v1")

  # Select a project by name
  gigwa_set_project("Nelson_et_al_2011")
  
  # Select a specific run by name
  gigwa_set_run("run1")
  
  # Get the list of all samples in the selected project
  samples <- gigwa_get_samples()

  # Get the list of all sequences in the selected project
  chroms <- gigwa_get_sequences()
  
  # Get the marker matrix
  geno_data <- gigwa_get_allelematrix(samples = samples[1:5],
                                      start   = 0,
                                      end     = 1234567,
                                      chrom   = chroms[1:3])
}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
