% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{gigwa_get_metadata}
\alias{gigwa_get_metadata}
\title{Get the Metadata of the Current Active GIGWA Run}
\usage{
gigwa_get_metadata()
}
\description{
This function is deprecated and will be removed in v2.0, use \code{\link{germplasm_get_metadata}} instead.
}
