% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{gigwa_list_dbs}
\alias{gigwa_list_dbs}
\title{List GIGWA Databases}
\usage{
gigwa_list_dbs()
}
\value{
A list of existing databases.
}
\description{
Get the list of existing databases in the current GIGWA server.
}
\examples{
if (interactive()) {
  # Configure your GIGWA connection
  set_qbms_config("https://gigwa.southgreen.fr/gigwa/", 
                  time_out = 300, engine = "gigwa", no_auth = TRUE)

  # List existing databases in the GIGWA server
  gigwa_list_dbs()
}
}
\seealso{
\code{\link{set_qbms_config}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
