% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gigwa.R
\name{gigwa_list_projects}
\alias{gigwa_list_projects}
\title{Get the List of All Projects in the Selected GIGWA Database}
\usage{
gigwa_list_projects()
}
\value{
A list of project names in the selected database.
}
\description{
Retrieve the list of projects available in the currently active GIGWA database, set
using `gigwa_set_db()`. If no database is selected, the function will throw an error.
}
\examples{
if (interactive()) {
  set_qbms_config("https://gigwa.southgreen.fr/gigwa/", 
                  time_out = 300, engine = "gigwa", no_auth = TRUE)
  gigwa_set_db("Sorghum-JGI_v1")
  gigwa_list_projects()
}
}
\seealso{
\code{\link{set_qbms_config}}, \code{\link{gigwa_set_db}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
