% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{gigwa_list_runs}
\alias{gigwa_list_runs}
\title{Get the List of the Run Names Available in the Selected GIGWA Project}
\usage{
gigwa_list_runs()
}
\description{
This function is deprecated and will be removed in v2.0, use \code{\link{genotyping_list_runs}} instead.
}
