% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{gigwa_set_db}
\alias{gigwa_set_db}
\title{Set the current active GIGWA database by name}
\usage{
gigwa_set_db(db_name)
}
\arguments{
\item{db_name}{the name of the database}
}
\value{
no return value
}
\description{
This function will update the current active database in the internal
configuration object (including the brapi connection object).
}
\examples{
if(interactive()) {
# config your GIGWA connection
set_qbms_config("https://gigwa.southgreen.fr/gigwa/", 
                time_out = 300, engine = "gigwa", no_auth = TRUE)

# select a database by name
gigwa_set_db("Sorghum-JGI_v1")
}
}
\seealso{
\code{\link{set_qbms_config}}, \code{\link{gigwa_list_dbs}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
