% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{ini_terraclimate}
\alias{ini_terraclimate}
\title{from <- '2019-09-01'
to   <- '2022-06-30'}
\usage{
ini_terraclimate(
  from = "2019-09-01",
  to = "2022-06-30",
  clim_vars = c("ppt", "tmin", "tmax"),
  data_path = "./data/",
  timeout = 300
)
}
\description{
clim_vars <- c('ppt', 'tmin', 'tmax')
data_path <- './data/'
}
