% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{list_crops}
\alias{list_crops}
\title{Get the List of Supported Crops}
\usage{
list_crops()
}
\value{
A list of supported crops.
}
\description{
Retrieves the list of supported crops.
}
\examples{
if(interactive()) {
  # Configure your BMS connection
  set_qbms_config("https://bms.icarda.org/ibpworkbench")

  # Login using your BMS account (interactive mode)
  # You can pass the BMS username and password as parameters (batch mode)
  login_bms()

  # List supported crops in the BMS server
  list_crops()
}

}
\seealso{
\code{\link{login_bms}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
