% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{list_locations}
\alias{list_locations}
\title{Get the list of locations information of the current selected crop}
\usage{
list_locations()
}
\value{
a data frame of the locations information
}
\description{
This function will retrieve the locations information of the current active crop
as configured in the internal state object using `set_crop()` function.
}
\seealso{
\code{\link{login_bms}}, \code{\link{set_crop}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
