% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gigwa.R
\name{list_variantsets}
\alias{list_variantsets}
\title{Get the List of the Variant Sets Available in the Selected Study}
\usage{
list_variantsets()
}
\value{
A list of set names associated with the selected study.
}
\description{
Retrieve the list of available variant sets in the currently active study, 
set using `set_study()`. If no study is selected, an error will be raised.
}
\seealso{
\code{\link{set_qbms_config}}, \code{\link{set_study}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
