% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{login}
\alias{login}
\title{Login to a BrAPI DB Server}
\usage{
login(username = NULL, password = NULL, ...)
}
\arguments{
\item{username}{The username (optional, default is NULL).}

\item{password}{The password (optional, default is NULL).}

\item{...}{Additional arguments passed to \code{\link{login_oauth2}}.}
}
\value{
No return value. The access token is stored internally for future use.
}
\description{
Logs in to the DB server using a username and password. If credentials are not provided,
a pop-up window will prompt the user. This function wraps around all `login_*` functions
and accepts additional parameters via `...`
}
\seealso{
\code{\link{login_oauth2}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
