% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{login_breedbase}
\alias{login_breedbase}
\title{Login to the BreedBase Server}
\usage{
login_breedbase(username = NULL, password = NULL)
}
\arguments{
\item{username}{The username (optional, default is NULL).}

\item{password}{The password (optional, default is NULL).}
}
\value{
No return value.
}
\description{
Logs in to the BreedBase server.
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
