% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{login_germinate}
\alias{login_germinate}
\title{Login to the Germinate Server}
\usage{
login_germinate(username = NULL, password = NULL)
}
\arguments{
\item{username}{The username (optional, default is NULL).}

\item{password}{The password (optional, default is NULL).}
}
\value{
No return value. The access token is stored internally for future use.
}
\description{
Login to the Germinate Server
}
\author{
Khaled Al-Shamaa (\email{k.el-shamaa@cgiar.org})
}
