% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{rbindx}
\alias{rbindx}
\title{Combine Data Frames by Row, Filling in Missing Columns}
\usage{
rbindx(..., dfs = list(...))
}
\arguments{
\item{...}{The first argument data frame.}

\item{dfs}{Input data frames to row bind together.}
}
\value{
A single data frame.
}
\description{
Combines a list of data frames by row, filling in missing columns with NA.
}
