% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{rstudio_is_available}
\alias{rstudio_is_available}
\title{Check if RStudio is Available}
\usage{
rstudio_is_available()
}
\value{
Logical TRUE if running in RStudio, FALSE otherwise.
}
\description{
Check if RStudio is Available
}
\keyword{internal}
