% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{search_program_germplasm}
\alias{search_program_germplasm}
\title{Search germplasm for a given program}
\usage{
search_program_germplasm(string = "INB841", type_of_search = "STARTSWITH")
}
\arguments{
\item{string}{string to search}

\item{type_of_search}{c("STARTSWITH", "CONTAINS") "STARTSWITH" by default}
}
\value{
data.frame with the search results
}
\description{
Search germplasm for a given program
}
\examples{
# In progress
}
