% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{set_crop}
\alias{set_crop}
\title{Set the Current Active Crop}
\usage{
set_crop(crop_name)
}
\arguments{
\item{crop_name}{The name of the crop.}
}
\value{
No return value.
}
\description{
Updates the current active crop in the internal configuration object (including 
the BrAPI connection object).
}
\examples{
if(interactive()) {
  # Configure your server connection
  set_qbms_config("https://bms.icarda.org/ibpworkbench")

  # Login using your account (interactive mode)
  # You can pass your username and password as parameters (batch mode)
  login_bms()

  # Select a crop by name
  set_crop("wheat")
}

}
\seealso{
\code{\link{login_bms}}, \code{\link{list_crops}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
