% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gigwa.R
\name{set_run}
\alias{set_run}
\title{Set the Current Active Variant Set}
\usage{
set_run(run_name)
}
\arguments{
\item{set_name}{The name of the variant set to be active.}
}
\value{
No return value. Updates the internal state with the selected variant set.
}
\description{
Select a variant set from the active study and set it as the current active 
variant set in the internal state, enabling further data retrieval operations.
}
\seealso{
\code{\link{set_qbms_config}}, \code{\link{set_study}}, \code{\link{list_runs}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
