% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gigwa.R
\name{set_variantset}
\alias{set_variantset}
\title{Set the Active Variant Set}
\usage{
set_variantset(variantset_name)
}
\arguments{
\item{variantset_name}{The name of the variant set to be active.}
}
\value{
No return value. Updates the internal state with the selected variant set.
}
\description{
Selects a variant set by name from the currently active study and updates the 
internal state to make it the active variant set. This selection is required 
for subsequent data retrieval operations related to variant sets.
}
\seealso{
\code{\link{set_qbms_config}}, \code{\link{set_study}}, \code{\link{list_variantsets}}
}
\author{
Khaled Al-Shamaa (\email{k.el-shamaa@cgiar.org})
}
