% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{brapi_get_allelematrix}
\alias{brapi_get_allelematrix}
\title{BrAPI Get Allele Matrix}
\usage{
brapi_get_allelematrix(
  samples = NULL,
  start = 0,
  end = "",
  chrom = NULL,
  snps = NULL,
  snps_pageSize = 10000,
  samples_pageSize = 100
)
}
\description{
BrAPI Get Allele Matrix
}
