% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{brapi_get_allelematrix}
\alias{brapi_get_allelematrix}
\title{BrAPI Get Allele Matrix}
\usage{
brapi_get_allelematrix(
  samples = NULL,
  start = 0,
  end = "",
  chrom = NULL,
  snps = NULL,
  snps_pageSize = 10000,
  samples_pageSize = 100,
  simplify = TRUE
)
}
\arguments{
\item{samples}{samples}

\item{start}{start}

\item{end}{end}

\item{chrom}{chrom}

\item{snps}{snps}

\item{snps_pageSize}{snps_pageSize}

\item{samples_pageSize}{samples_pageSize}

\item{simplify}{simplify}
}
\description{
BrAPI Get Allele Matrix
}
\examples{
if (interactive()) {
  # Configure your GIGWA connection
  set_qbms_config("https://gigwa.southgreen.fr/gigwa/", 
                  time_out = 300, engine = "gigwa", no_auth = TRUE)

  # Select a database by name
  gigwa_set_db("Sorghum-JGI_v1")

  # Select a project by name
  gigwa_set_project("Nelson_et_al_2011")
  
  # Select a specific run by name
  gigwa_set_run("run1")
  
  # Get a list of all samples in the selected run
  samples <- gigwa_get_samples()
  
  chk <- Sys.time()
  marker_matrix <- gigwa_get_variants(start = 0, 
                                      end = 012345678, 
                                      referenceName = "Sb01", 
                                      samples = samples)
  round(Sys.time()-chk, 2)
  
  chk <- Sys.time()
  geno_data <- brapi_get_allelematrix(samples = samples, 
                                      start = 0, 
                                      end = 012345678, 
                                      chrom = "Sb01", 
                                      snps = marker_matrix$`rs#`) 
  round(Sys.time()-chk, 2)
  
  all.equal(geno_data, marker_matrix[,5:12], check.attributes = FALSE)
}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
