% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{brapi_get_call}
\alias{brapi_get_call}
\title{Internal Function Used for Core BrAPI GET Calls}
\usage{
brapi_get_call(call_url, nested = TRUE)
}
\arguments{
\item{call_url}{BrAPI URL to call in GET method.}

\item{nested}{Logical indicating whether retrieved JSON data will be flattened. Default is TRUE.}
}
\value{
Result object returned by the JSON API response.
}
\description{
This function is created for *internal use only* to call BrAPI in the GET method and
retrieve the raw response data and send back the results. This function takes
care of pagination, authentication, encoding, compression, decoding JSON response, etc.
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
