% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi.R
\docType{data}
\name{brapi_map}
\alias{brapi_map}
\title{BrAPI Mapping Table}
\format{
An object of class \code{data.frame} with 34 rows and 3 columns.
}
\usage{
brapi_map
}
\description{
This internal table maps QBMS function names to corresponding BrAPI calls for different versions 
of the BrAPI standard. It is used internally by the QBMS system to translate function calls into 
BrAPI-compliant API requests.
}
\details{
This table supports both BrAPI v1 and v2 endpoints. It is regularly updated to reflect changes 
in the BrAPI standard and the inclusion of new API calls.
}
\keyword{datasets}
