% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\docType{data}
\name{brapi_map}
\alias{brapi_map}
\title{/brapi/{brapi_ver}/{brapi_call}}
\format{
An object of class \code{data.frame} with 28 rows and 3 columns.

An object of class \code{data.frame} with 28 rows and 3 columns.

An object of class \code{data.frame} with 28 rows and 3 columns.

An object of class \code{data.frame} with 28 rows and 3 columns.

An object of class \code{data.frame} with 28 rows and 3 columns.

An object of class \code{data.frame} with 28 rows and 3 columns.
}
\usage{
brapi_map

brapi_map

brapi_map

brapi_map

brapi_map

brapi_map
}
\description{
/brapi/{brapi_ver}/{brapi_call}

need to be deprecated in favor of list_dbs() and set_db() functions
https://github.com/plantbreeding/BrAPI/issues/495

POST: germplasmDbIds, observationLevel = "PLOT"

POST: studyDbIds

POST: studyDbIds

POST: germplasmDbIds
}
\keyword{datasets}
