% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.R
\name{build_pedigree_table}
\alias{build_pedigree_table}
\title{Building Pedigree Table Recursively}
\usage{
build_pedigree_table(
  geno_list = NULL,
  pedigree_list = NULL,
  pedigree_df = NULL
)
}
\arguments{
\item{geno_list}{A list of genotypes/germplasm names.}

\item{pedigree_list}{A list of associated pedigree strings.}

\item{pedigree_df}{Pedigree data.frame as per the previous call/iteration.}
}
\value{
A data.frame with three columns corresponding to the identifiers for the 
individual, female parent, and male parent, respectively.
}
\description{
Internal helper function to build the pedigree table recursively.
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
