% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{build_pedigree_table}
\alias{build_pedigree_table}
\title{Building Pedigree Table Recursively}
\usage{
build_pedigree_table(
  geno_list = NULL,
  pedigree_list = NULL,
  pedigree_df = NULL
)
}
\arguments{
\item{geno_list}{List of genotypes/germplasms names.}

\item{pedigree_list}{List of associated pedigree strings.}

\item{pedigree_df}{Pedigree data.frame as per previous call/iteration.}
}
\value{
A data.frame that has three columns correspond to the identifiers for the individual,
        female parent and male parent, respectively.
}
\description{
Internal helping function to build the pedigree table recursively.
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
