% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{debug_qbms}
\alias{debug_qbms}
\title{Debug internal QBMS status object}
\usage{
debug_qbms()
}
\description{
Return the internal QBMS status object for debuging
}
\examples{
View(debug_qbms())
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
