% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi.R
\name{debug_qbms}
\alias{debug_qbms}
\title{Debug Internal QBMS Status Object}
\usage{
debug_qbms()
}
\value{
An environment object containing package configuration and status.
}
\description{
Returns the internal QBMS status object for debugging purposes.
}
\examples{
obj <- debug_qbms()
obj$config
obj$state

}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
