% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{get_async_page}
\alias{get_async_page}
\title{Asynchronously Fetch a Single API Page}
\usage{
get_async_page(full_url, nested)
}
\arguments{
\item{full_url}{Character string specifying the full URL of the API endpoint to request.}

\item{nested}{Logical value indicating whether to flatten nested lists in the JSON response.}
}
\value{
A future representing the asynchronous operation, which will resolve to a list containing the parsed JSON response.
}
\description{
Sends an asynchronous HTTP GET request to fetch data from a single API page.
}
\details{
This function uses \code{future::future()} to perform the HTTP GET request asynchronously.
It retrieves the content from the specified URL, checks for HTTP errors, and parses the JSON response.
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
