% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{get_async_page}
\alias{get_async_page}
\title{Async version of HTTP GET request}
\usage{
get_async_page(full_url, nested)
}
\value{
Async version of HTTP GET request.
}
\description{
A small helper function to create `async` version of the original HTTP GET request.
}
