% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{get_async_page}
\alias{get_async_page}
\title{Async Version of HTTP GET Request}
\usage{
get_async_page(full_url, nested)
}
\arguments{
\item{full_url}{URL to retrieve.}

\item{nested}{Logical indicating whether to flatten nested data frames. Default is FALSE.}
}
\value{
Async version of the HTTP GET request.
}
\description{
A small helper function to create an `async` version of the original HTTP GET request.
}
