% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{get_async_pages}
\alias{get_async_pages}
\title{Asynchronously Fetch Multiple API Pages}
\usage{
get_async_pages(pages, nested)
}
\arguments{
\item{pages}{Character vector of full URLs specifying the API endpoints to request.}

\item{nested}{Logical value indicating whether to flatten nested lists in the JSON responses.}
}
\value{
A list of parsed JSON responses from each page.
}
\description{
Sends asynchronous HTTP GET requests to fetch data from multiple API pages concurrently.
}
\details{
This function uses \code{future.apply::future_lapply()} to perform concurrent HTTP GET requests for multiple pages.
It retrieves and parses the JSON responses from each URL provided.
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
