% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{get_async_pages}
\alias{get_async_pages}
\title{Run all supplied pages}
\usage{
get_async_pages(pages, nested)
}
\arguments{
\item{pages}{List of URLs to retrieve}

\item{nested}{Don't flatten nested data.frames}
}
\value{
Async deferred object.
}
\description{
A small helper function to create a deferred value that is resolved when all 
listed pages are resolved.
}
