% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{get_async_pages}
\alias{get_async_pages}
\title{Run All Supplied Pages}
\usage{
get_async_pages(pages, nested)
}
\arguments{
\item{pages}{List of URLs to retrieve.}

\item{nested}{Logical indicating whether to flatten nested data frames. Default is FALSE.}
}
\value{
Async deferred object.
}
\description{
A small helper function to create a deferred value that is resolved when all 
listed pages are resolved.
}
