% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi.R
\name{get_brapi_url}
\alias{get_brapi_url}
\title{Get the BrAPI Endpoint URL for a given QBMS function}
\usage{
get_brapi_url(func_name)
}
\arguments{
\item{func_name}{(string) Function name}
}
\value{
The BrAPI endpoint URL to call
}
\description{
Get the BrAPI Endpoint URL for a given QBMS function
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
