% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{get_germplasm_attributes}
\alias{get_germplasm_attributes}
\title{Get germplasm attributes for a given germplasmDbId in a crop}
\usage{
get_germplasm_attributes(
  germplasmDbId = "49f77560-7874-11eb-91c9-0242ac140003"
)
}
\arguments{
\item{germplasmDbId}{id in the database. It can be more than one.}
}
\value{
data.frame with attributes
}
\description{
Get germplasm attributes for a given germplasmDbId in a crop
}
\examples{
# library(QBMS)
# 
# # config your BMS connection (by providing your BMS login page URL)
# set_qbms_config("https://bms.ciat.cgiar.org/ibpworkbench/controller/auth/login")
# 
# # login using your BMS account (interactive mode)
# # or pass your BMS username and password as parameters (batch mode)
# login_bms()
# 
# # list supported crops in the current bms server
# list_crops()
# 
# # select a crop by name
# set_crop("bean")
# 
# # list all breeding programs in the selected crop
# list_programs()
# 
# # select a breeding program by name
# set_program("CIAT-Mesoamerican beans")
# 
# results <- search_program_germplasm(string = "SER21", type_of_search = "STARTSWITH")
# ids <- results$germplasmUUID[1:2]
# 
# get_germplasm_attributes(germplasmDbId = ids)
}
