% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{get_germplasm_data}
\alias{get_germplasm_data}
\title{Get the observations data of a given germplasm name}
\usage{
get_germplasm_data(germplasm_name)
}
\arguments{
\item{germplasm_name}{the name of the germplasm}
}
\value{
a data frame of the germplasm observations data aggregate from all trials
}
\description{
This function will retrieve the observations data of the current active study
as configured in the internal state object using `set_study()` function.
}
\examples{
# config your BMS connection
set_qbms_config("https://bms.example.com/ibpworkbench/controller/auth/login")

# login using your BMS account (interactive mode)
# you can pass BMS username and password as parameters (batch mode)
login_bms()

set_crop("Tutorial1")

# select a breeding program by name
set_program("Training Breeding Program")

# retrive observations data of a given germplasm aggregated from all trials
germplasm_observations <- get_germplasm_data("FLIP10-3C")
}
\seealso{
\code{\link{login_bms}}, \code{\link{set_crop}}, \code{\link{set_program}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
