% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{get_germplasm_id}
\alias{get_germplasm_id}
\title{Get Germplasm ID}
\usage{
get_germplasm_id(germplasm_name = "")
}
\arguments{
\item{germplasm_name}{The name of the germplasm.}
}
\value{
A string of the germplasm ID.
}
\description{
Retrieves the germplasm ID for the given germplasm name in the current crop.
}
\seealso{
\code{\link{set_crop}}, \code{\link{get_germplasm_data}}, \code{\link{get_germplasm_attributes}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
