% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{get_parents}
\alias{get_parents}
\title{Get Direct Parents}
\usage{
get_parents(pedigree)
}
\arguments{
\item{pedigree}{String provide the parentage through which a cultivar was obtained.}
}
\value{
Vector of two items, the direct female and male parents.
}
\description{
Internal helping function to split the given pedigree string that provides the parentage
through which a cultivar was obtained, and get the pedigrees of the direct parents.
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
