% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{get_parents}
\alias{get_parents}
\title{Get Direct Parents}
\usage{
get_parents(pedigree)
}
\arguments{
\item{pedigree}{A string providing the parentage through which a cultivar was obtained.}
}
\value{
A vector of two items, representing the direct female and male parents.
}
\description{
Internal helper function to split the given pedigree string that provides the parentage
through which a cultivar was obtained and retrieve the pedigrees of the direct parents.
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
