% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{get_terraclimate}
\alias{get_terraclimate}
\title{Get TerraClimate data for a given location}
\usage{
get_terraclimate(
  lat,
  lon,
  from = "1958-01-01",
  to = "2020-12-31",
  clim_vars = NULL,
  month_mask = NULL
)
}
\arguments{
\item{lat}{Latitude in decimal degree format.}

\item{lon}{Longitude in decimal degree format.}

\item{from}{Start date as a string in the 'YYYY-MM-DD' format.}

\item{to}{End date as a string in the 'YYYY-MM-DD' format.}

\item{clim_vars}{List of all climate variables to be imported. Valid list includes: \emph{aet, def, pet,
ppt, q, soil, srad, swe, tmax, tmin, vap, ws, vpd, and PDSI}. Default is NULL for all.}

\item{month_mask}{A list of all months of interest (e.g., planting season: \code{c(10:12,1:5)}). 
Default is NULL for all.}
}
\value{
A list of two data.frame(s):
\enumerate{
\item \emph{climate:} includes the climate variables (\href{https://www.climatologylab.org/terraclimate-variables.html}{reference}).
\item \emph{biovars:} includes the calculated bioclimatic variables (\href{https://www.worldclim.org/data/bioclim.html}{reference}).
}
}
\description{
\enumerate{
\item \href{https://www.nature.com/articles/sdata2017191}{Scientific Paper}
\item \href{https://www.climatologylab.org/terraclimate.html}{About the Dataset}
\item \href{http://thredds.northwestknowledge.net:8080/thredds/terraclimate_aggregated.html}{Hosting Catalog}
\item \href{https://hpc.uidaho.edu/}{Service Provider}
\item \href{https://www.climatologylab.org/terraclimate-variables.html}{Climate Variables}
\item \href{https://www.worldclim.org/data/bioclim.html}{Bioclimatic Vars}
\item \href{https://github.com/rspatial/dismo/blob/master/R/biovars.R}{Calculating biovars}
}

\strong{TerraClimate vs. \href{https://www.worldclim.org/data/worldclim21.html}{WorldClim}}
\itemize{
\item 1958-2020 vs. 1970-2000
\item 14 vs. 7 climate variables
\item ~4 km vs. ~1 km spatial resolution
\item need to calculate vs. pre-calculated 19 bioclimatic variables
}
}
\examples{
if(interactive()) {
# data <- get_terraclimate(36.016, 36.943, '1979-09-01', '2012-06-30', c('ppt', 'tmin', 'tmax'), c(10:12,1:5))
data <- get_terraclimate(36.016, 36.943, '1979-09-01', '2012-06-30')

View(data$climate)

View(data$biovars)
}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
