% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{gigwa_get_samples}
\alias{gigwa_get_samples}
\title{Get the Samples List of the Current Active GIGWA Run}
\usage{
gigwa_get_samples()
}
\description{
This function is deprecated and will be removed in v2.0, use \code{\link{genotyping_list_samples}} instead.
}
