% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{gigwa_get_variants}
\alias{gigwa_get_variants}
\title{marker_matrix <- gigwa_get_variants(max_missing = 0.2, min_maf = 0.35, samples = c("ind1", "ind3", "ind7"))}
\usage{
gigwa_get_variants(max_missing = 1, min_maf = 0, samples = NULL)
}
\description{
marker_matrix <- gigwa_get_variants(max_missing = 0.2, min_maf = 0.35, samples = c("ind1", "ind3", "ind7"))
}
