% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{gigwa_set_run}
\alias{gigwa_set_run}
\title{Set the Current Active GIGWA Run}
\usage{
gigwa_set_run(run_name)
}
\arguments{
\item{run_name}{run name}
}
\description{
This function is deprecated and will be removed in v2.0, use \code{\link{genotyping_set_run}} instead.
}
