% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{list_crops}
\alias{list_crops}
\title{Get the list of supported crops}
\usage{
list_crops()
}
\value{
a list of supported crops
}
\description{
Get the list of supported crops
}
\examples{
if(interactive()) {
# config your BMS connection
set_qbms_config("https://www.bms-uat-test.net/ibpworkbench")

# login using your BMS account (interactive mode)
# you can pass BMS username and password as parameters (batch mode)
login_bms()

# list supported crops in the bms server
list_crops()
}
}
\seealso{
\code{\link{login_bms}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
