% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{list_locations}
\alias{list_locations}
\title{Get the List of Locations Information of the Current Selected Crop}
\usage{
list_locations()
}
\value{
A data frame containing information about locations relevant to the current crop.
}
\description{
Retrieves a list of locations associated with the current active crop, as 
configured in the internal state object using the \code{\link{set_crop}} function.
}
\seealso{
\code{\link{login_bms}}, \code{\link{set_crop}} for related crop operations.
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
