% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{list_locations}
\alias{list_locations}
\title{Get the List of Locations Information of the Current Selected Crop}
\usage{
list_locations()
}
\value{
A data frame of the locations information.
}
\description{
Retrieves the locations information of the current active crop as configured 
in the internal state object using the `set_crop()` function.
}
\seealso{
\code{\link{login_bms}}, \code{\link{set_crop}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
