% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{list_programs}
\alias{list_programs}
\title{Get the list of breeding programs names}
\usage{
list_programs()
}
\value{
a list of breeding programs names
}
\description{
This function will retrieve the breeding programs list from the current active 
crop as configured in the internal configuration object using `set_crop()`
function.
}
\examples{
if (interactive()) {
# config your BMS connection
set_qbms_config("https://www.bms-uat-test.net/ibpworkbench")

# login using your BMS account (interactive mode)
# you can pass BMS username and password as parameters (batch mode)
login_bms()

set_crop("maize")

# list existing breeding programs
list_programs()
}
}
\seealso{
\code{\link{login_bms}}, \code{\link{set_crop}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
