% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gigwa.R
\name{list_variantsets}
\alias{list_variantsets}
\title{List Variant Sets in the Selected Study}
\usage{
list_variantsets()
}
\value{
A list of the names of variant sets associated with the selected study.
}
\description{
Retrieves the names of all variant sets available in the currently selected 
study, as set by \code{set_study()}. If no study is selected, the function returns 
an error.
}
\seealso{
\code{\link{set_qbms_config}}, \code{\link{set_study}}
}
\author{
Khaled Al-Shamaa (\email{k.el-shamaa@cgiar.org})
}
