% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{login}
\alias{login}
\title{Login to a BrAPI Server}
\usage{
login(username = NULL, password = NULL, ...)
}
\arguments{
\item{username}{The username (optional, default is NULL).}

\item{password}{The password (optional, default is NULL).}

\item{...}{Additional arguments passed to \code{\link{login_oauth2}}.}
}
\value{
No return value. On success, the access token is stored internally for future use.
}
\description{
Authenticates with a BrAPI server using a username and password. If credentials 
are not provided, interactive prompt will request them from the user. This 
function serves as a wrapper for all \code{login_*} authentication methods and 
accepts additional parameters via \code{...}
}
\seealso{
\code{\link{login_oauth2}}
}
\author{
Khaled Al-Shamaa (\email{k.el-shamaa@cgiar.org})
}
