% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{login_bms}
\alias{login_bms}
\title{Login to the BMS server}
\usage{
login_bms(username = NULL, password = NULL)
}
\arguments{
\item{username}{the BMS username (optional, default is NULL)}

\item{password}{the BMS password (optional, default is NULL)}
}
\description{
Connect to the BMS server. If username or password parameters are missing, 
then a login window will pop-up to insert username and password. 

All other connection parameters (i.e. server IP or domain, connection port, 
API path, and connection protocol e.g. http://) will retrieve from the 
qbms_config list.

This function will update both of the qbms_config list (brapi connection 
object in the con key) and qbms_state list (token value in the token key).
}
\examples{
# config your BMS connection
set_qbms_config("https://bms.example.com/ibpworkbench/controller/auth/login")

# login using your BMS account (interactive mode)
# you can pass BMS username and password as parameters (batch mode)
login_bms()
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
