% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{login_breedbase}
\alias{login_breedbase}
\title{Login to the BreedBase server}
\usage{
login_breedbase(username = NULL, password = NULL)
}
\arguments{
\item{username}{the username (optional, default is NULL)}

\item{password}{the password (optional, default is NULL)}
}
\value{
no return value
}
\description{
Login to the BreedBase server
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
