% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{login_breedbase}
\alias{login_breedbase}
\title{Login to the BreedBase Server}
\usage{
login_breedbase(username = NULL, password = NULL)
}
\arguments{
\item{username}{The username (optional, default is NULL).}

\item{password}{The password (optional, default is NULL).}
}
\value{
No return value. The access token is stored internally for future use.
}
\description{
Logs in to the BreedBase server using a username and password. If credentials are not provided,
a pop-up window will prompt the user. The function is a wrapper around the `login_bms()` function,
with encoding set to `form`.
}
\examples{
if(interactive()) {
  set_qbms_config("https://cassavabase.org/", engine = "breedbase")
  
  # Login using your BreedBase account (interactive mode)
  login_breedbase()
  
   # You can pass BreedBase username and password as parameters (batch mode)
   # login_breedbase("username", "password")
}

}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
