% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{login_gigwa}
\alias{login_gigwa}
\title{set_qbms_config("https://gigwa.southgreen.fr/gigwa/", time_out = 300, engine = "gigwa", no_auth = TRUE)
login_gigwa("gigwadmin", "nimda")}
\usage{
login_gigwa(username = NULL, password = NULL)
}
\description{
set_qbms_config("https://gigwa.southgreen.fr/gigwa/", time_out = 300, engine = "gigwa", no_auth = TRUE)
login_gigwa("gigwadmin", "nimda")
}
