% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\docType{data}
\name{qbms_globals}
\alias{qbms_globals}
\title{List of non-BrAPI calls in QBMS functions}
\format{
An object of class \code{environment} of length 2.
}
\usage{
qbms_globals
}
\description{
get_program_studies()
/crops/{cropName}/programs/{programUUID}/studies/{studyId}/entries/metadata (BMS: get study entries metadata)
}
\details{
get_germplasm_list()
/crops/{cropName}/programs/{programUUID}/studies/{studyId}/entries (BMS: get entry type) (POST: body = "")

get_trial_obs_ontology()
to get the observationVariableNames list
/crops/{cropName}/variables/filter?programUUID={programUUID}&variableIds={id1,id2,..} (BMS: gets all variables using filter)

gigwa_get_variants()
/ga4gh/variants/search 

dancing steps: 
- searchMode = 0 to get total
- then searchMode = 3 to request actual results
- keep checking progress status /gigwa/progress
- then call the same /ga4gh/variants/search to get the ready results

GA4GH: https://rest.ensembl.org/documentation/info/gavariants
BrAPI: https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.1#/Allele%20Matrix
Internal state variables/lists
}
\keyword{datasets}
