% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\docType{data}
\name{qbms_globals}
\alias{qbms_globals}
\title{Name:     qbms.R
Purpose:  Set of functions to query BMS by a wrapper using BrAPI calls
Author:   Khaled Al-Shamaa <k.el-shamaa@cgiar.org>
Version:  0.7.0}
\format{
An object of class \code{environment} of length 2.
}
\usage{
qbms_globals
}
\description{
License:  GPLv3
Load/install required packages
if (!require(utils)) install.packages("utils")
if (!require(httr)) install.packages("httr")
if (!require(tcltk)) install.packages("tcltk")
if (!require(jsonlite)) install.packages("jsonlite")
Internal state variables/lists
}
\keyword{datasets}
