% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{rbindlistx}
\alias{rbindlistx}
\title{Make One Data.Table from a List of Many}
\usage{
rbindlistx(x)
}
\arguments{
\item{x}{A list containing data.table, data.frame, or list objects.}
}
\value{
An unkeyed data.table containing a concatenation of all the items passed in.
}
\description{
Performs the equivalent of do.call("rbind", x) on data.frames, but much faster.
}
