% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{rbindx}
\alias{rbindx}
\title{Combine data.frames by row, filling in missing columns}
\usage{
rbindx(..., dfs = list(...))
}
\arguments{
\item{...}{input data frames to row bind together.}
}
\value{
a single data frame
}
\description{
rbinds a list of data frames filling missing columns with NA
}
