% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{scan_brapi_endpoints}
\alias{scan_brapi_endpoints}
\title{Scan BrAPI endpoints}
\usage{
scan_brapi_endpoints(programDbId = 0, trialDbId = 0, studyDbId = 0)
}
\arguments{
\item{programDbId}{programDbId used for BrAPI endpoints scanning (default is 0)}

\item{trialDbId}{trialDbId used for BrAPI endpoints scanning (default is 0)}

\item{studyDbId}{studyDbId used for BrAPI endpoints scanning (default is 0)}
}
\value{
a data.frame list the QBMS function, BrAPI endpoint URL, and status
}
\description{
Scan available BrAPI endpoints in the configured source server
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
