% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi.R
\name{scan_brapi_endpoints}
\alias{scan_brapi_endpoints}
\title{Scan BrAPI Endpoints}
\usage{
scan_brapi_endpoints(programDbId = 0, trialDbId = 0, studyDbId = 0)
}
\arguments{
\item{programDbId}{(numeric) The programDbId to scan specific program-related endpoints (default is 0).}

\item{trialDbId}{(numeric) The trialDbId to scan specific trial-related endpoints (default is 0).}

\item{studyDbId}{(numeric) The studyDbId to scan specific study-related endpoints (default is 0).}
}
\value{
A data frame listing the QBMS function, BrAPI endpoint URL, and availability status for each endpoint.
}
\description{
Scans the available BrAPI endpoints on the configured source server and checks their accessibility. 
This function allows users to verify which BrAPI endpoints are available based on the provided IDs.
}
\author{
Khaled Al-Shamaa (\email{k.el-shamaa@cgiar.org})
}
