% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{set_crop}
\alias{set_crop}
\title{Set the current active crop}
\usage{
set_crop(crop_name)
}
\arguments{
\item{crop_name}{the name of the crop}
}
\value{
no return value
}
\description{
This function will update the current active crop in the internal 
configuration object (including the brapi connection object).
}
\examples{
if(interactive()){
# config your BMS connection
set_qbms_config("https://www.bms-uat-test.net/ibpworkbench")

# login using your BMS account (interactive mode)
# you can pass BMS username and password as parameters (batch mode)
login_bms()

set_crop("maize")
}
}
\seealso{
\code{\link{login_bms}}, \code{\link{list_crops}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
