% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{set_crop}
\alias{set_crop}
\title{Set the current active crop}
\usage{
set_crop(crop_name)
}
\description{
This function will update the current active crop in the internal 
configuration object (including the brapi connection object).
}
\examples{
# config your BMS connection
set_qbms_config("https://bms.example.com/ibpworkbench/controller/auth/login")

# login using your BMS account (interactive mode)
# you can pass BMS username and password as parameters (batch mode)
login_bms()

set_crop("Tutorial1")
}
\seealso{
\code{\link{login_bms}}, \code{\link{list_crops}}
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
