% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{set_qbms_config}
\alias{set_qbms_config}
\title{Configure BMS server settings}
\usage{
set_qbms_config(
  url = "http://localhost/ibpworkbench/controller/auth/login",
  path = "bmsapi",
  page_size = 1000,
  time_out = 120,
  no_auth = FALSE,
  engine = "bms"
)
}
\arguments{
\item{url}{URL of the BMS login page (default is "http://localhost/ibpworkbench/")}

\item{path}{BMS API path (default is "bmsapi")}

\item{page_size}{Page size (default is 1000)}

\item{time_out}{Number of seconds to wait for a response until giving up (default is 10)}

\item{no_auth}{TRUE if the server doesn't require authentication/login (default is FALSE)}

\item{engine}{Backend database (qbms default, breedbase)}
}
\value{
no return value
}
\description{
Set the connection configuration of the BMS server
}
\examples{
set_qbms_config("https://www.bms-uat-test.net/ibpworkbench")
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
