% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbms.R
\name{set_qbms_config}
\alias{set_qbms_config}
\title{Configure BMS server settings}
\usage{
set_qbms_config(
  server = "localhost",
  port = 80,
  protocol = "http://",
  path = "bmsapi",
  page_size = 1000
)
}
\arguments{
\item{server}{BMS server domain name or IP address (default is "localhost")}

\item{port}{Connection port (default is 48080)}

\item{protocol}{Connection protocol (default is "http://")}

\item{path}{BMS API path (default is "bmsapi")}

\item{page_size}{Page size (default is 1000)}
}
\description{
Set the connection configuration of the BMS server
}
\examples{
set_qbms_config(server = "bms.icarda.org", port = 18443, protocol = "https://")
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
