% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{set_token}
\alias{set_token}
\title{Set Access Token Response}
\usage{
set_token(token, user = "", expires_in = 3600)
}
\arguments{
\item{token}{The access token string issued by the authorization server.}

\item{user}{The username associated with the token (optional).}

\item{expires_in}{The lifetime of the access token in seconds (optional, default is 3600 seconds).}
}
\value{
No return value. Updates the internal state with the token info.
}
\description{
Stores the access token and associated details (such as username and expiration time) in the internal state.
The token is typically retrieved from the server during login and used for subsequent API requests.
}
\author{
Khaled Al-Shamaa, \email{k.el-shamaa@cgiar.org}
}
